# Improve Store::getInProperty performance, refs 1234