--- a/src/Defines.php
+++ b/src/Defines.php
@@ -203,3 +203,10 @@ define( 'SMW_LINV_OBFU', 4 ); // Using the Obfuscator approach
 define( 'SMW_RF_NONE', 0 );
 define( 'SMW_RF_TEMPLATE_OUTSEP', 2 ); // #2022 Enable 2.5 behaviour for template handling
 /**@}*/
+
+/**@{
+  * Constants for $smwgEntityLookupFeatures
+  */
+define( 'SMW_EL_NONE', 0 );
+define( 'SMW_EL_INPROP', 2 ); // New query for EntityLookup::getInProperties
+/**@}*/