--- a/src/SQLStore/EntityStore/DIHandlers/DITimeHandler.php
+++ b/src/SQLStore/EntityStore/DIHandlers/DITimeHandler.php
@@ -47,6 +47,24 @@ class DITimeHandler extends DataItemHandler {
 	 *
 	 * {@inheritDoc}
 	 */
+	public function getTableIndexes() {
+		return array(
+
+			// SMWSQLStore3Readers::fetchSemanticData
+			// SELECT p.smw_title as prop,o_serialized AS v0, o_sortkey AS v2
+			// FROM `smw_di_time` INNER JOIN `smw_object_ids` AS p ON
+			// p_id=p.smw_id WHERE s_id='104822'	7.9291ms
+			// ... FROM `smw_fpt_sobj` INNER JOIN `smw_object_ids` AS o0 ON
+			// o_id=o0.smw_id WHERE s_id='104322'
+			's_id,p_id,o_sortkey,o_serialized',
+		);
+	}
+
+	/**
+	 * @since 1.8
+	 *
+	 * {@inheritDoc}
+	 */
 	public function getWhereConds( DataItem $dataItem ) {
 		return array( 'o_sortkey' => $dataItem->getSortKey() );
 	}