--- a/src/SQLStore/EntityStore/DIHandlers/DIWikiPageHandler.php
+++ b/src/SQLStore/EntityStore/DIHandlers/DIWikiPageHandler.php
@@ -52,7 +52,46 @@ class DIWikiPageHandler extends DataItemHandler {
 	 * {@inheritDoc}
 	 */
 	public function getTableIndexes() {
-		return array( 'o_id' );
+		return array(
+			'o_id',
+
+			// SMWSQLStore3Readers::fetchSemanticData
+			// ... FROM `smw_fpt_sobj` INNER JOIN `smw_object_ids` AS o0 ON
+			// o_id=o0.smw_id WHERE s_id='104322'
+			's_id,o_id',
+
+			// SMWSQLStore3Readers::fetchSemanticData
+			// ... FROM `smw_di_wikipage` INNER JOIN `smw_object_ids` AS p ON
+			// p_id=p.smw_id INNER JOIN `smw_object_ids` AS o0 ON o_id=o0.smw_id
+			// WHERE s_id='104815'
+			's_id,p_id,o_id',
+
+			// QueryEngine::getInstanceQueryResult
+			// ... INNER JOIN `smw_fpt_inst` AS t3 ON t2.smw_id=t3.s_id WHERE
+			// (t2.smw_namespace='0' AND (t3.o_id='56')
+			'o_id,s_id',
+
+			// QueryEngine::getInstanceQueryResult
+			//'p_id,o_id,s_sort',
+
+			// SMWSQLStore3Readers::getPropertySubjects
+			//'p_id,s_sort,s_id',
+
+			// SMWSQLStore3Readers::getPropertySubjects
+			// SELECT DISTINCT s_id FROM `smw_fpt_sobj` ORDER BY s_sort
+			//'s_sort,s_id',
+
+			// SELECT DISTINCT s_id FROM `smw_fpt_subp` WHERE o_id='96' ORDER BY s_sort
+			//'o_id,s_sort,s_id',
+
+			// In-property lookup
+			'o_id,p_id',
+			//'o_id,p_id,s_sort',
+
+			// SMWSQLStore3Readers::getPropertySubjects
+			// SELECT DISTINCT s_id FROM `smw_di_wikipage` WHERE (p_id='64' AND o_id='104') ORDER BY s_sort ASC
+			//'o_id,p_id,s_id,s_sort'
+		);
 	}
 
 	/**