--- a/src/SQLStore/TableBuilder/PostgresTableBuilder.php
+++ b/src/SQLStore/TableBuilder/PostgresTableBuilder.php
@@ -286,7 +286,7 @@ EOT;
 		}
 
 		$tableName = $this->connection->tableName( $tableName, 'raw' );
-		$indexName = "{$tableName}_index{$indexName}";
+		$indexName = $this->getCumulatedIndexName( $tableName, $columns );
 
 		$this->reportMessage( "   ... creating new index $columns ..." );
 
@@ -295,6 +295,12 @@ EOT;
 		}
 
 		$this->reportMessage( "done.\n" );
+	}
+
+	private function getCumulatedIndexName( $tableName, $columns ) {
+		// Identifiers -- table names, column names, constraint names,
+		// etc. -- are limited to a maximum length of 63 bytes
+		return str_replace( '__' , '_', "{$tableName}_idx_" . str_replace( array( '_', 'smw', ',' ), array( '', '_', '_' ), $columns ) );
 	}
 
 	private function getIndexInfo( $tableName ) {