This PR is made in reference to: #1234

This PR addresses or contains:

- The following [content](https://gist.github.com/mwjames/103afa6ecd53b666de075e69cc1970ca) was used to simulate some high data load
- Changes on pre/post to the query plan can be found [here](https://gist.github.com/mwjames/0bbe876b11ec4e134cddd0c46345b7d5)
- The PR in particular addresses the `Store::getInProperties` performance with a __significant improvement__ in query execution time by extending the `INNER JOIN` with a subquery to pre-select and filter possible matches before being materialized. One example entity with > 100 incoming properties and an overall table size that covers ~1M triples demonstrated the following changes pre/post this PR.
  - MySQL ( `841.3639ms` ↘ `39.1431ms`)
  - Postgres (`1282.6319ms` ↘ `59.8922ms`) 
- The suggestion from #1234 on using something like  ` ... WHERE smw_id IN (SELECT p_id FROM mediawiki."smw_di_wikipage" WHERE o_id='19683’);` did not create the expected improvement on MySQL hence a different query was selected that showed favorable query plans on both MySQL and Postgres.

To eliminate filesort and other expensive operations we needed to:

* Create additional composite indexes to ensure the query planner choose the right one with the objective of using an index
* Eliminate `ORDER BY` as it is expected to returns all matches and sort the results during a post processing
*  Filter possible matches before being materialized

```sql
SELECT DISTINCT smw_title,smw_sortkey,smw_iw FROM "smw_object_ids"
INNER JOIN (
  SELECT p_id FROM "smw_di_wikipage" WHERE o_id='266' GROUP BY p_id
 ) AS t1 ON t1.p_id=smw_id
WHERE smw_iw!=':smw' AND smw_iw!=':smw-delete'
```

Tested on:

```
MediaWiki	1.30.0-alpha (3d26a96)
PHP	7.1.1 (apache2handler)
MariaDB	10.1.21-MariaDB
ICU	57.1
```
```
MediaWiki	1.26.2
PHP	5.6.8 (apache2handler)
PostgreSQL	9.3.12
ICU	54.1
```

This PR includes:
- [x] Tests (unit/integration)
- [x] CI build passed