@mwjames 2017-05-14
This requires either `update.php` or `setupStore.php` to be run.

@kghbln One thing to be noted is due to the added indices, the `update.php` might take a bit longer the first time as those are created and depend largely on the size of the DB.