@kghbln 2017-05-14
@mwjames This surely will come out with a bang! The performance improvements seen are quite impressive! WOW!

> One thing to be noted is due to the added indices, the update.php might take a bit longer the first time as those are created and depend largely on the size of the DB.

We should note this is the RELEASE NOTES in any case.