@mwjames 2017-07-11
Well, the setting is internal in case something seems fishy in order
to easily switch back and compare results without reworking the whole
code from the inside.

I hope that by the time of SMW 3.0 the setting can vanish and yes, the
default value is different.

On 7/12/17, Karsten Hoffmeyer <notifications@github.com> wrote:
> kghbln commented on this pull request.
>
>
>
>> @@ -1495,4 +1495,18 @@
>  	'smwgEntityCollation' => 'identity',
>  	##
>
> +	##
> +	# Entity lookup specific features
> +	#
> +	# - SMW_EL_NONE applies no query or schema changes
> +	#
> +	# - SMW_EL_INPROP enables a new query form for selecting incoming
> properties
> +	#   (#1234)
> +	#
> +	# @since 3.0
> +	# @default false
>
> @mwjames Shouldn't this say `SMW_EL_INPROP`
>
> --
> You are receiving this because you were mentioned.
> Reply to this email directly or view it on GitHub:
> https://github.com/SemanticMediaWiki/SemanticMediaWiki/pull/2461#pullrequestreview-49240406