@kghbln 2017-07-11
> I hope that by the time of SMW 3.0 the setting can vanish and yes, the default value is different.

You mean the whole `smwgEntityLookupFeatures` parameter or just "false" as the internal setting?