@kghbln 2017-10-13
[Documented on wiki](https://www.semantic-mediawiki.org/wiki/Help:$smwgEntityLookupFeatures)